Decide to make slight modifications to the objects architecture
The GameManager object will be a central role and in the future
it will communicate with both the MainActivity as well as the
SharedPreferences. Therefore, instead of a linear architecture,
the nodes Prefer. and MainAct. will conjoin in the GameManager
node. Although not yet implemented, the generateNewConf()
function in the GameManager will return a MinesweeperAdapter,
which the MainActivity will use to set the RecyclerView.