Set the OnTileClick() function in GameManager under major reconstruction
Decided on an architecture between the GameManager and Rules class
The GameManager will communicate directly with the Rules class,
will communicate the input types, the coordinates of the currently
selected tile. On the other hand, the Rules class will return
specific outputs through its static functions that will decide
overall the flow of the game. Commented the OnTileClick() body.