Implement uncover_empty_tile() function, run tests succesfully
I have implemented the first part of the uncovering architecture,
more specifically the function that uncovers when an empty tile is
detonated. For searching those tiles which are to be uncovered, a
queue has been used along with a vector of objects. Besides
implementing this feature, I have fixed a minor bug regarding the
second part of the uncovering architecture. That bug has persisted
in the doesItHaveEnoughFlagsSet() function and was attributed to
the misplacement of the tileIcon and tileValue variables.